requirements.txt
pandas==2.3.2
python-telegram-bot==22.4
black==25.9.0
pylint==3.3.8